Game act: Add functions to switch to a new game act.
Ending the current game act and starting a new one will be driven by
a Lua script in a dialog or an event.

The actual switch can not be executed while the Lua VM is running, we
have to wait to be returned in the main loop of the game, because we
have to free all game data (even memory used by dialogs and events)
before to start the new game act.

Consequently, a first function is called by the Lua script, recording
the fact that the current game act is finished. Then, an other function
will be called in the main loop to executed the switch.