Reorder passive skills, include Animal Magnetism based on patch by Miles
Passive Skills are meant to be at bottom from skill list, not at middle.

Animal Magnetism is a passive skill which increase length of takeover game
in 0.5 seconds per level.
This does not includes a source book because this, just like other passive skills,
should be limited to NPC's quests.

To test it, it's currently necessary to cheat.
There will be a quest for this later.