Change a variable name to better clarify its functionality.
Variable "is_a_living" is used in droid_specs.lua as a marker for
terminals to limit drawing of extra rotation images. Variable name
is misleading and is changed with this patch to highly that a terminal
does not move. This is consistent with usage in droid_specs.lua file.
Only TRM and STM have this variable set to false (no other droid specs
call this variable).
The default value defined in luaconfig.c is for value to be true.