Swap the color channels of the thumbnail images if needed.
Before c180a255, make_texture_out_of_surface()'s first step was to
convert the loaded surface to a POT surface.
During that convertion, an SDL blit is called, that possibly swaps
the color channels of the surface to suit the framebuffer
configuration.

With c180a255, that POT convertion is no more called.

It seemed to have no visual impact, because the function that loads an image
from the disk (load_image_surface()) also swaps the color channels if needed
(through the call to SDL_DisplayFormatAlpha()).

However, thumbnails are not loaded using that function, and so the color
channels are not swapped. This patch changes the function that loads a
thumbnail image to mimic load_image_surface().