Specialize bots being used as boss.
for short, places the guardians from DSB, MSO and HFF on their own droidspec.
They have much more hitpoints and experience reward, but weapons are the same.

This will give more feeling of fighting a boss than an actual challenge.

The 598 Minister (Boss) received a small change to recover time, too.
Boss have lots of takeover charges.
It won't be easy to get them to fight with you!