Remove resolutions lower than 640x480 from available resolutions, and reformat the information messages displayed when a wrong resolution is requested.
Some panels needs at least a 640x480 screen, and the game quits with a
fatal error if the screen is smaller.

Some menus (such as the load game and the delete game menus) are also
too bigs to suit a screen smaller than 640x480.

Based on a patch from Gregory Lozet.