The Master of Melee: Erin, the Bothunter
We have someone to help Tux to attain maximum proficiency on Ranged,
but not on Programming and Melee.

This patch fixes the proficiency problems on Melee by adding a new character
near the Hell Fortress: Erin.
Of course her base model got a weapon...

Erin have background story with Saul (see key characters section on website)
and with Iris too.
In future there'll be a sidequest here for Rebel Faction.
Actually there is a quest to retrieve a misterious necklace that
no one knows from where it came for,
what purpose it serves or if it's merely of sentimental value.