Separating Ammo Shop from Equipment Shop (Lukas)
Basically this is done in a rather hack-ish and unheard-of way.
All ammo was entrusted to Benjamin but instead of allowing him to sell,
I added another node to Lukas. So you buy Benjamin's ammo from Lukas.

Pistol sale intensity was reduced on Lukas.
Exterminator Ammo sale intensity was increased.

Ammo for weapons which cannot be found (normally) in game are NOT SOLD.