Dialog C engine: remove a problematic extraneous lua_pop()
When cheat keys are enabled, one can exit a running dialog as soon as a
user interaction is awaited.
If a coroutine was interrupted by a lua_yield, then the coroutine must
be removed from the lua stack.
However, no lua_pop() has to be called if no coroutine is running.

The current code was lua_pop-ing in all cases, possibly letting the lua stack
in a bad state.