Add 'event timers'.
An 'event timer' is an event scheduled to be executed at a later time.

Like any event, it is defined in events.dat. It has no specific type,
and only defines the Lua code to be executed (if the trigger is
enabled).

A Lua function can be used, in a dialog or an other event, to schedule
the delayed event: dispatch_event("<event_name>", <delay_in_seconds>)

Event timers still to be executed are saved in the savegame.